fix: fixed paths and login

+++ src/Controllers/UsersController.php
@@ -18,23 +18,35 @@ class UsersController
 
     public function store()
     {
+        // Väldi header viga – ära väljasta midagi enne seda
+        ob_start();
+
         $user = new User();
         $user->email = $_POST['email'];
         $user->password = password_hash($_POST['password'], PASSWORD_DEFAULT);
         $user->save();
 
+        ob_end_clean();
         header('Location: /users');
+        exit;
     }
 
     public function edit()
     {
-        $id = $_GET['id'];
+        $id = $_GET['id'] ?? null;
+        if (!$id) {
+            header('Location: /users');
+            exit;
+        }
+
         $user = User::find($id);
         include __DIR__ . '/../../views/users/edit.php';
     }
 
     public function update()
     {
+        ob_start();
+
         $id = $_POST['id'];
         $user = User::find($id);
         $user->email = $_POST['email'];
@@ -43,15 +55,23 @@ class UsersController
         }
         $user->save();
 
+        ob_end_clean();
         header('Location: /users');
+        exit;
     }
 
     public function destroy()
     {
+        ob_start();
+
         $id = $_POST['id'];
         $user = User::find($id);
-        $user->delete();
+        if ($user) {
+            $user->delete();
+        }
 
+        ob_end_clean();
         header('Location: /users');
+        exit;
     }
 }
